requirements.txt
google-genai
streamlit==1.22.0
PyPDF2
json